Update Java version and Gradle distribution
Updated APP_JAVA_VERSION from 17 to 20 in .github/workflows/gradle_ci.yml. This is to keep up with the latest JDK. Also, simplified the cache action and condensed the condition for the dropbox job. Updated the Gradle wrapper distribution URL in gradle-wrapper.properties for the same reason of staying up-to-date. Lastly, adjusted the source and target compatibility in build.gradle.kts to reflect updated Java version.